+++ modules/registration_module/registration.py
@@ -32,6 +32,7 @@ class Perform(gf.Page):
         self.channel_position = QLineEdit(placeholderText='eg. 0 (default) / 1 / ...')
         self.channel_position.setMinimumWidth(200)
         self.channel_position.setValidator(QIntValidator())
+        self.channel_position.setText("0")
 
         self.use_input_folder = QRadioButton("Use input image folder")
         self.use_input_folder.setChecked(True)