+++ modules/segmentation_module/segmentation.py
@@ -1,8 +1,8 @@
 import logging
 import os
-from PyQt5.QtWidgets import QFileDialog, QCheckBox, QPushButton, QVBoxLayout, QHBoxLayout, QWidget, QGroupBox, QRadioButton, QApplication, QSpinBox, QFormLayout, QLabel, QLineEdit
+from PyQt5.QtWidgets import QFileDialog, QCheckBox, QPushButton, QVBoxLayout, QHBoxLayout, QWidget, QGroupBox, QRadioButton, QApplication, QSpinBox, QFormLayout, QLabel, QLineEdit, QComboBox
 from PyQt5.QtCore import Qt, QRegExp
-from PyQt5.QtGui import QCursor, QRegExpValidator
+from PyQt5.QtGui import QCursor, QIntValidator, QRegExpValidator
 from modules.segmentation_module import segmentation_functions as f
 from general import general_functions as gf
 import torch
@@ -44,6 +44,52 @@ class Segmentation(QWidget):
         self.output_filename_label.setFrame(False)
         self.output_filename_label.setEnabled(False)
 
+        self.channel_position = QLineEdit(placeholderText='eg. 0 (default) / 1 / ...')
+        self.channel_position.setMinimumWidth(200)
+        self.channel_position.setValidator(QIntValidator())
+        self.channel_position.setText("0")
+        # Z-Projection range
+        # only bestZ
+        self.projection_mode_bestZ = QRadioButton("Z section with best focus")
+        self.projection_mode_bestZ.setChecked(False)
+        self.projection_mode_bestZ.setToolTip('Keep only Z section with best focus.')
+        # around bestZ
+        self.projection_mode_around_bestZ = QRadioButton("Range around Z section with best focus")
+        self.projection_mode_around_bestZ.setChecked(True)
+        self.projection_mode_around_bestZ.setToolTip('Project all Z sections with Z in the interval [bestZ-range,bestZ+range], where bestZ is the Z section with best focus.')
+        self.projection_mode_around_bestZ_zrange = QSpinBox()
+        self.projection_mode_around_bestZ_zrange.setMinimum(0)
+        self.projection_mode_around_bestZ_zrange.setMaximum(20)
+        self.projection_mode_around_bestZ_zrange.setValue(3)
+        # fixed range
+        self.projection_mode_fixed = QRadioButton("Fixed range")
+        self.projection_mode_fixed.setChecked(False)
+        self.projection_mode_fixed.setToolTip('Project all Z sections with Z in the interval [from,to].')
+        self.projection_mode_fixed_zmin = QSpinBox()
+        self.projection_mode_fixed_zmin.setMinimum(0)
+        self.projection_mode_fixed_zmin.setMaximum(20)
+        self.projection_mode_fixed_zmin.setValue(4)
+        self.projection_mode_fixed_zmin.valueChanged.connect(self.projection_mode_fixed_zmin_changed)
+        self.projection_mode_fixed_zmax = QSpinBox()
+        self.projection_mode_fixed_zmax.setMinimum(0)
+        self.projection_mode_fixed_zmax.setMaximum(20)
+        self.projection_mode_fixed_zmax.setValue(6)
+        self.projection_mode_fixed_zmax.valueChanged.connect(self.projection_mode_fixed_zmax_changed)
+        # all
+        self.projection_mode_all = QRadioButton("All Z sections")
+        self.projection_mode_all.setChecked(False)
+        self.projection_mode_all.setToolTip('Project all Z sections.')
+        # Z-Projection type
+        self.projection_type = QComboBox(self)
+        self.projection_type.addItem("max")
+        self.projection_type.addItem("min")
+        self.projection_type.addItem("mean")
+        self.projection_type.addItem("median")
+        self.projection_type.addItem("std")
+        self.projection_type.setCurrentText("mean")
+        self.projection_type.setDisabled(self.projection_mode_bestZ.isChecked())
+        self.projection_mode_bestZ.toggled.connect(self.projection_type.setDisabled)
+
         self.use_gpu = QCheckBox("Activate GPU")
         self.use_gpu.setChecked(False)
         self.coarse_grain = QCheckBox("Activate coarse grain parallelisation")
@@ -70,6 +116,7 @@ class Segmentation(QWidget):
         layout2.addWidget(self.image_list)
         groupbox.setLayout(layout2)
         layout.addWidget(groupbox)
+        
         groupbox = QGroupBox("Cellpose model")
         layout2 = QVBoxLayout()
         layout3 = QHBoxLayout()
@@ -78,6 +125,7 @@ class Segmentation(QWidget):
         layout2.addLayout(layout3)
         groupbox.setLayout(layout2)
         layout.addWidget(groupbox)
+        
         groupbox = QGroupBox("Output")
         layout2 = QVBoxLayout()
         layout2.addWidget(QLabel("Folder:"))
@@ -101,6 +149,55 @@ class Segmentation(QWidget):
         layout2.addLayout(layout3)
         groupbox.setLayout(layout2)
         layout.addWidget(groupbox)
+
+        groupbox = QGroupBox("Options")
+        layout3 = QFormLayout()
+        layout3.setLabelAlignment(Qt.AlignLeft)
+        layout3.setFormAlignment(Qt.AlignLeft)
+        groupbox2 = QGroupBox("If multiple channels:")
+        layout4 = QFormLayout()
+        layout4.addRow("Channel position:",self.channel_position)
+        groupbox2.setLayout(layout4)
+        layout3.addRow(groupbox2)
+
+        groupbox2 = QGroupBox("If multiple z:")
+        layout4 = QFormLayout()
+        # Z-Projection range
+        widget = QWidget()
+        layout5 = QVBoxLayout()
+        layout5.addWidget(self.projection_mode_bestZ)
+        layout5.addWidget(self.projection_mode_around_bestZ)
+        groupbox3 = QGroupBox()
+        groupbox3.setToolTip('Project all Z sections with Z in the interval [bestZ-range,bestZ+range], where bestZ is the Z section with best focus.')
+        groupbox3.setVisible(self.projection_mode_around_bestZ.isChecked())
+        self.projection_mode_around_bestZ.toggled.connect(groupbox3.setVisible)
+        layout6 = QFormLayout()
+        layout6.addRow("Range:",self.projection_mode_around_bestZ_zrange)
+        groupbox3.setLayout(layout6)
+        layout5.addWidget(groupbox3)
+        layout5.addWidget(self.projection_mode_fixed)
+        groupbox3 = QGroupBox()
+        groupbox3.setToolTip('Project all Z sections with Z in the interval [from,to].')
+        groupbox3.setVisible(self.projection_mode_fixed.isChecked())
+        self.projection_mode_fixed.toggled.connect(groupbox3.setVisible)
+        layout6 = QHBoxLayout()
+        layout7 = QFormLayout()
+        layout7.addRow("From:",self.projection_mode_fixed_zmin)
+        layout6.addLayout(layout7)
+        layout7 = QFormLayout()
+        layout7.addRow("To:",self.projection_mode_fixed_zmax)
+        layout6.addLayout(layout7)
+        groupbox3.setLayout(layout6)
+        layout5.addWidget(groupbox3)
+        layout5.addWidget(self.projection_mode_all)
+        widget.setLayout(layout5)
+        layout4.addRow("Projection range:",widget)
+        layout4.addRow("Projection type:",self.projection_type)
+        groupbox2.setLayout(layout4)
+        layout3.addRow(groupbox2)
+        groupbox.setLayout(layout3)
+        layout.addWidget(groupbox)
+
         groupbox = QGroupBox("Multi-processing")
         layout2 = QVBoxLayout()
         layout2.addWidget(self.use_gpu)
@@ -147,6 +244,14 @@ class Segmentation(QWidget):
 
         self.output_filename_label.setText(os.path.join(output_path,"<input basename>" + self.output_suffix + self.output_user_suffix.text() + ".ome.tif"))
 
+    def projection_mode_fixed_zmin_changed(self, value):
+        if self.projection_mode_fixed_zmax.value() < value:
+            self.projection_mode_fixed_zmax.setValue(value)
+
+    def projection_mode_fixed_zmax_changed(self, value):
+        if self.projection_mode_fixed_zmin.value() > value:
+            self.projection_mode_fixed_zmin.setValue(value)
+
     def submit(self):
         """
         Retrieve the input parameters
@@ -178,6 +283,20 @@ class Segmentation(QWidget):
                 self.logger.error('More than one input file will output to the same file (output files will be overwritten).\nEither use input image folder as output folder or avoid processing images from different input folders.\nProblematic input files:\n%s', '\n'.join(duplicates[:4] + (['...'] if len(duplicates) > 4 else [])))
                 return False
             return True
+
+        if self.channel_position.text() == '':
+            channel_position = 0
+        else:
+            channel_position = int(self.channel_position.text())
+        projection_type = self.projection_type.currentText()
+        if self.projection_mode_bestZ.isChecked():
+            projection_zrange = 0
+        elif self.projection_mode_around_bestZ.isChecked():
+            projection_zrange = self.projection_mode_around_bestZ_zrange.value()
+        elif self.projection_mode_fixed.isChecked():
+            projection_zrange = (self.projection_mode_fixed_zmin.value(), self.projection_mode_fixed_zmax.value())
+        elif self.projection_mode_all.isChecked():
+            projection_zrange = None
 
         image_paths = self.image_list.get_file_list()
         model_path = self.selected_model.text()
@@ -220,6 +339,9 @@ class Segmentation(QWidget):
                  model_path,
                  output_path,
                  output_basename,
+                 channel_position,
+                 projection_type,
+                 projection_zrange,
                  n_count,
                  self.display_results.isChecked(),
                  self.use_gpu.isChecked()